Ai_assistant-local-voice/requirements.txt
ollama
python-dotenv
SpeechRecognition
websocket-client
RealtimeTTS
RealtimeSTT
faster-whisper
torch
edge-tts
pyaudio
fastapi
uvicorn
pynput
requests
